v1.0.6 seperated quiz logic and UI
